Create City-of-Boulder-Colorado-USA
The purpose of Boulder Open Data Catalog is to increase public access to high value, machine readable datasets generated by the City of Boulder. We invite you to actively participate in shaping the future of Boulder's Open Data Catalog by contacting us to suggest additional datasets or building and submitting applications with the available data.